Fix calculation of pageoffset when relocating index entries.
The previous formula had no chance of being correct if the number
of pages per segment was >1, and even then it might not have
worked if we'd trimmed a number of index entries straight out
of the metapage that was not a multiple of the number that will
fit in an index page. Try to do better.

Along the way, fix an off-by-one error in a sanity check.

Per test case from Dilip Kumar.