Vacuum dead item store over the conveyor belt
This provides a mechanism for storing and retrieving dead items into the
conveyor belt.  Basically, this is an interface using which a vacuum
module can interact with the conveyor belt.  This maintains a DeadItemState
which keeps track of the current insertion progress i.e the first and the
last conveyor belt page for the current vacuum run.  And on the completion
of the vacuum run, this takes care of setting the complete vacuum run
bound by storing the last conveyor belt pageno of the current vacuum run
into the special space of the first conveyor belt page for this run.

This also provides the infrastructure to avoid adding duplicate items into
the conveyor belt.  Basically, if we perform the first vacuum pass multiple
times without executing the second vacuum pass then it is possible that we
encounter the same dead items in the conveyor belt so this module maintains
a cache over the conveyor belt such that it only loads the data into the
cache w.r.t the current block the vacuum is processing so we don't need to
maintain a huge cache.