Decoupling table and index vacuum
Currently, the table vacuum and index vacuum are executed as a single
operation.  Basically, we vacuum the table by performing hot pruning
and remembering dead items in the cache and then we perform the index
vacuum and perform the second pass of the heap vacuum under that we
mark items unused.  But, the problem with this is that every index has
its own needs, which could be separate from the needs of the tables.

In this patch, we make these multiple vacuum passes as independent
operations.  So the idea that we provide multiple vacuum options
under that the user can perform the independent operations
i.e. "VACUUM (heap_hot_prune) tbl_name" for performing just the
hot prune (first vacuum) pass, "VACUUM (heap_vacuum) tbl_name"
for the second heap pass to set dead item unused for which index
vacuum is done.  And additionally, we are now allowing users to just
perform the index vacuum i.e. "VACUUM idx_name".

So under the heap_hot_prune pass, we will generate the dead tids and
instead of directly performing the index vacuum we will flush those
dead tids to the conveyor belt using Deadtidstore interfaces.  Then
in the index pass, we will read the data from the conveyor belt and
perform the index vacuum and at last, in the heap_vacuum pass, we
will read the data from the conveyor belt and mark all dead items
unused.  However, in the second pass we can only mark those items
unused which are dead, and for which all the indexes for the table are
already vacuumed.  So for identifying that in the pg_class entry we
store the conveyor belt pageno up to which we have already done the
index vacuum for the index related entry and we have already done the
heap_vacuum pass for the table related entry.