Conditionally skip the index vacuum
During vacuum first heap pass currently, if we do not have
enough space in the maintenance work mem during the first
pass of the heap then we pause and perform the index and heap
second pass vacuuming and throw away the memory dead items and
continue with the first heap pass.  As part of this patch, we skip
the index vacuum for all or some of the indexes based on the indexam
skip routine.  So if we need to do that then we must store the
current dead items in some permanent storage.  So as part of this
patch, whenever the dead item cache is full that time we are flushing
those dead items to the conveyor belt.  And once the first pass is
complete then we will conditionally perform the index and heap
vacuuming.